requirements.txt
flask
gunicorn
textblob
google.generativeai